backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "bcrypt==4.3.0",
    "cffi==1.17.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography==44.0.2",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email_validator==2.2.0",
    "fastapi==0.104.1",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "mysql-connector-python==8.0.33",
    "passlib==1.7.4",
    "protobuf==3.20.3",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "pydantic_core==2.27.2",
    "PyJWT==2.10.1",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "rsa==4.9",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.15",
    "starlette==0.27.0",
    "typing_extensions==4.12.2",
    "uvicorn==0.21.1",
    "websockets==11.0.2"
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "ARG001",
]
ignore = [
    "E501",
    "B008",
    "W191",
    "B904",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
